requirements.txt
pydantic>=1.10
PyYAML>=6.0
pytest>=7.0
scikit-learn>=1.3.0
joblib>=1.3.0
mlflow>=2.6.0
redis>=3.5
fastapi==0.95.2
uvicorn[standard]==0.21.1
httpx==0.24.0
